+++ spec/rails_app/app/components/grid/filters_off.rb
@@ -1,7 +1,0 @@
-class Grid::FiltersOff < Netzke::Basepack::Grid
-  def configure(c)
-    super
-    c.model = "Book"
-    c.enable_column_filters = false
-  end
-end