+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.11.1 - WIP
 * Fix filtering by association when associtaion method is not string
+* Fix decimal columns in advanced search form
 
 # 0.11.0 - 2015-02-05
 * Rails 4.2